src/log_collectors/gcp_audit_logs.py
"""
GCP Audit Logs Collector for Chronicle SIEM
Collects and processes GCP audit logs for security analysis
"""

from google.cloud import audit_logs_v1
from google.cloud import storage
from typing import Dict, List

class GCPAuditLogCollector:
    def __init__(self, project_id: str, destination_bucket: str):
        self.project_id = project_id
        self.destination_bucket = destination_bucket
        self.client = audit_logs_v1.AuditLogsClient()
        self.storage_client = storage.Client()

    def setup_log_sink(self):
        """Configure log sink for audit logs to Cloud Storage"""
        sink_name = f"chronicle-audit-logs-{self.project_id}"
        destination = f"storage.googleapis.com/{self.destination_bucket}"
        
        sink = {
            "name": sink_name,
            "destination": destination,
            "filter": 'resource.type="gce_instance" OR resource.type="cloudsql_database"'
        }
        return self.client.create_sink(parent=f"projects/{self.project_id}", sink=sink)

    def process_logs(self, event: Dict) -> Dict:
        """Process and enrich audit log entries"""
        enriched_event = {
            "timestamp": event.get("timestamp"),
            "resource": event.get("resource", {}),
            "principal_email": event.get("protoPayload", {}).get("authenticationInfo", {}).get("principalEmail"),
            "method_name": event.get("protoPayload", {}).get("methodName"),
            "severity": event.get("severity"),
            "project_id": self.project_id
        }
        return enriched_event

    def export_to_chronicle(self, events: List[Dict]):
        """Export processed logs to Chronicle SIEM"""
        # Implement Chronicle ingestion logic here
        pass
