requirements.txt
requirements.txt
google-cloud-audit-logs==2.0.0
google-cloud-storage==2.0.0
google-cloud-vpc-access==1.0.0
google-cloud-compute==2.0.0
google-cloud-security-center==1.0.0
google-cloud-secret-manager==2.0.0
google-cloud-bigquery==2.0.0
google-cloud-pubsub==2.0.0
google-cloud-dlp==3.0.0
functions-framework==3.0.0
requests==2.26.0
pandas==1.3.0
numpy==1.21.0
tensorflow==2.6.0
scikit-learn==0.24.2
python-dateutil==2.8.2
pyyaml==5.4.1
prometheus-client==0.11.0
